% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllpointscornerCombined.R
\docType{data}
\name{AllpointscornerCombined}
\alias{AllpointscornerCombined}
\title{All Points Corner Combined}
\format{
\subsection{\code{who$fac}}{

A data frame with 193 rows and 3 columns:
\describe{
\item{Name}{outline name}
\item{Type}{Projectile point type}
\item{Origin}{column created from file name format--no relevance to analysis}
}
}
}
\source{
see table 1
}
\usage{
AllpointscornerCombined
}
\description{
This is a Momocs Open object containing all corner notched projectile points used in the publication. The object consists of a list of 'coo' objects, which are the closed outline objects, and a fac object which is a data frame of metadata. The fac is described below.
}
\keyword{datasets}
