% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllpointsOutlines.R
\docType{data}
\name{AllpointsOutlines}
\alias{AllpointsOutlines}
\title{All Points Outlines}
\format{
\subsection{\code{who$fac}}{

A data frame with 242 rows and 3 columns:
\describe{
\item{Name}{outline name}
\item{Origin}{origin site of the projectile point}
\item{Type}{Projectile point type}
}
}
}
\source{
see table 1
}
\usage{
AllpointsOutlines
}
\description{
This is a Momocs Open object containing all projectile points used in the publication. The object consists of a list of 'coo' objects, which are the closed outline objects, and a fac object which is a data frame of metadata. The fac is described below.
}
\keyword{datasets}
