% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FremontProjectilePointData.R
\docType{data}
\name{FremontProjectilePointData}
\alias{FremontProjectilePointData}
\title{Fremont Projectile Point Data}
\format{
\subsection{\code{FremontProjectilePointData}}{

A data frame with 23 rows and 6 columns:
\describe{
\item{Site}{Archaeological site}
\item{Period}{General time period used in publication}
\item{Total Projectile Points}{Sum of following two columns}
\item{Total Rosegate Points}{Sum of projectile points identified as Rosegate}
\item{Total Dart Points}{Sum of projectile points identified as archaic or associated with the Atlatl}
\item{References}{Origin of data}
}
}
}
\source{
see table 1
}
\usage{
FremontProjectilePointData
}
\description{
This is a data frame object containing metadata on the total number of projectile points found at each site.
}
\keyword{datasets}
