% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WolfVillageOutlines.R
\docType{data}
\name{WolfVillageOutlines}
\alias{WolfVillageOutlines}
\title{Wolf Village Outlines}
\format{
\subsection{\code{who$fac}}{

A data frame with 66 rows and 1 column:
\describe{
\item{Name}{The feature specimen number of the projectile point}
}
}
}
\source{
see table 1
}
\usage{
WolfVillageOutlines
}
\description{
This is a Momocs Open object containing all projectile points used in this publication from the site of Wolf Village. The object consists of a list of 'coo' objects, which are the closed outline objects, and a fac object which is a data frame of metadata. The fac is described below.
}
\keyword{datasets}
